+++ packages/unmock-cli/src/commands/index.ts
@@ -1,0 +1,3 @@
+export { init } from "./init";
+export { open } from "./open";
+export { list } from "./list";