+++ packages/unmock-core/src/interfaces.ts
@@ -560,7 +560,10 @@ export interface IListener {
 export interface IUnmockOptions extends ILogger {
   useInProduction(): boolean;
   isWhitelisted(url: string): boolean;
-  flaky(): boolean;
+  /**
+   * Randomize responses instead of using a fixed seed.
+   */
+  randomize(): boolean;
 }
 
 export interface IUnmockPackage {
@@ -671,7 +674,7 @@ export interface ISerializedResponse {
 
 export type CreateResponse = (
   request: ISerializedRequest,
-) => ISerializedResponse | undefined;
+) => ISerializedResponse;
 
 export type IStateTransformer = (
   req: ISerializedRequest,