+++ packages/unmock-fetch/src/fetch.ts
@@ -1,5 +1,6 @@
 import debug from "debug";
 import {
+  CreateResponse,
   ISerializedResponse,
   OnSerializedRequest,
 } from "unmock-core/dist/interfaces";
@@ -9,21 +10,34 @@ import { Response } from "./types";
 
 const debugLog = debug("unmock:fetch-mitm");
 
-export const buildFetch = (onSerializedRequest: OnSerializedRequest): Fetch =>
+const isCreateResponse = (
+  cb: CreateResponse | OnSerializedRequest,
+): cb is CreateResponse => cb.length === 1; // Check the number of function arguments
+
+export const buildFetch = (cb: CreateResponse | OnSerializedRequest): Fetch =>
   function fetch(url: RequestInfo, init?: RequestInit): Promise<Response> {
-    const req = serialize(url);
+    const req = serialize(url, init);
     debugLog(`Serialized request: ${JSON.stringify(req)}, init: ${init}`);
 
     return new Promise((resolve, reject) => {
       const sendResponse = (res: ISerializedResponse): void => {
-        const responseInit: ResponseInit = { status: res.statusCode };
+        const responseInit: ResponseInit = { status: res.statusCode }; // TODO Headers
         resolve(new Response(res.body, responseInit));
       };
 
       const emitError = (e: Error) => reject(e);
 
       setImmediate(() => {
-        onSerializedRequest(req, sendResponse, emitError);
+        if (isCreateResponse(cb)) {
+          try {
+            const res = cb(req);
+            sendResponse(res);
+          } catch (err) {
+            reject(err);
+          }
+        } else {
+          cb(req, sendResponse, emitError);
+        }
       });
     });
   };