+++ packages/unmock-node/src/backend.ts
@@ -1,5 +1,8 @@
 import { Backend } from "unmock-core";
-import { IInterceptorOptions } from "unmock-core/src/interceptor";
+import {
+  IInterceptorFactory,
+  IInterceptorOptions,
+} from "unmock-core/src/interceptor";
 import createFsServiceDefLoader from "./fs-service-def-loader";
 import NodeInterceptor from "./interceptor/node-interceptor";
 import FSLogger from "./loggers/filesystem-logger";
@@ -7,6 +10,7 @@ import FSSnapshotter from "./loggers/snapshotter";
 
 export interface INodeBackendOptions {
   servicesDirectory?: string;
+  interceptorFactory?: IInterceptorFactory;
 }
 
 /**
@@ -22,10 +26,12 @@ export default class NodeBackend extends Backend {
       }),
       FSSnapshotter.getOrUpdateSnapshotter({}),
     ];
+    const interceptorFactory =
+      (config && config.interceptorFactory) ||
+      ((options: IInterceptorOptions) => new NodeInterceptor(options));
     const serviceDefLoader = createFsServiceDefLoader(servicesDirectory);
     super({
-      interceptorFactory: (options: IInterceptorOptions) =>
-        new NodeInterceptor(options),
+      interceptorFactory,
       listeners,
       serviceDefLoader,
     });