+++ packages/unmock-server/index.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+const main = require("./dist/index.js");
+main.run();