+++ packages/unmock-server/src/constants.ts
@@ -1,0 +1,3 @@
+export const SERVER_HTTP_PORT = 8000;
+export const SERVER_HTTPS_PORT = 8443;
+export const PROXY_PORT = 8008;