+++ packages/unmock/src/__tests__/end-to-end/simple-service-with-runner.test.ts
@@ -29,6 +29,7 @@ beforeAll(() => {
   foo = unmock.on().services.foo;
 });
 afterAll(() => unmock.off());
+
 describe("Simple service test with runner", () => {
   it(
     "Should reset spies correctly",