+++ changelog.md
@@ -1,5 +1,13 @@
 # CHANGELOG.md
 
+## 0.3.10 (2019-11-4)
+
+Features:
+
+  - Add `unmock-server`, a mock server
+  - Add `unmock.randomize.on()` option for randomizing responses
+  - Introduce `unmock` CLI in `unmock-cli` package
+
 ## 0.3.9 (2019-10-17)
 
 Features: